requirements.txt
typer
tabulate
networkx
matplotlib
groq
flask
fpdf2